+++ redaudit/core/syn_scanner.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """
 RedAudit - SYN Scanner Module
-Copyright (C) 2025  Dorin Badea
+Copyright (C) 2026  Dorin Badea
 GPLv3 License
 
 v4.3: Optional SYN-based port scanning using scapy.